+++ app/stories/index.js
@@ -1,2 +1,5 @@
+import 'font-awesome/css/font-awesome.css'
 import './Welcome.js'
 import './ProjectChartStory.js'
+import './PagesStory.js'
+import './SearchStory.js'